% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_population}
\alias{get_population}
\title{Get population}
\usage{
get_population(method = "nearest_est")
}
\arguments{
\item{method}{"nearest_est" or "interpolate_proj"}
}
\description{
Get population
}
