% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{get_ta}
\alias{get_ta}
\title{TA matcher}
\usage{
get_ta(targ_ta, out_col = "ta_code")
}
\arguments{
\item{targ_ta}{Column of target TAs}

\item{out_col}{Name of TA details to extract}
}
\description{
Matches TA details
}
