% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{make_seasonal}
\alias{make_seasonal}
\title{Seasonal adjustment}
\usage{
make_seasonal(val, period)
}
\arguments{
\item{val}{Value column}

\item{period}{Period column}
}
\description{
Makes seasonally adjusted column from \link{period, value} dataframe.
}
\examples{
df \%>\% mutate(value_sadj = make_seasonal(value, period))
}
\keyword{clean}
\keyword{hud}
\keyword{seasonal}
\keyword{seasonality}
