% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{match_area}
\alias{match_area}
\title{Unknown/multi-level area matcher}
\usage{
match_area(src_area, out_col = "area_name")
}
\arguments{
\item{src_area}{Column of target areas}

\item{out_col}{Name of column to extract}
}
\description{
Matches region/TA/local board details by match_name,
without knowing what type of area it is.
}
