% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\docType{data}
\name{sample_wide}
\alias{sample_wide}
\title{Sample wide data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 183 rows and 3 columns.
}
\usage{
sample_wide
}
\description{
Jobs Online data in wide format
}
\keyword{datasets}
