% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{get_population}
\alias{get_population}
\title{Population provider}
\usage{
get_population(area, period, area_type)
}
\arguments{
\item{area}{Column of area}

\item{period}{Column of period, or a single period}

\item{area_type}{Type of area ("regc", "ta", "akl_board")}
}
\description{
For a given set of TA/region/local boards + period, provide monthly total population count.
Monthly data is interpolated from population estimates.
}
